+++ packages/onboarding/src/step-container-v2/components/TopBar/style.scss
@@ -34,6 +34,9 @@
 
 	&-right-button {
 		margin-left: auto;
+		--wp-components-color-accent: var(--color-neutral-100);
+		font-size: 0.875rem;
+		line-height: 1;
 	}
 
 	&-wordpress-logo {
