+++ packages/onboarding/src/step-container-v2/components/buttons/SkipButton/style.scss
@@ -1,6 +1,5 @@
 .step-container-v2__skip-button {
-	--wp-components-color-accent: var(--color-neutral-100);
-	font-size: 0.875rem;
-	line-height: 1;
+	font-size: inherit;
+	line-height: inherit;
 	font-weight: 500;
 }
